backend/api_gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "GraphQL API Gateway for shortening links"
authors = [
    { name = "neojelll", email = "timofejmitusin60892@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.12.0",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "strawberry-graphql[debug-server]>=0.258.0",
]

[project.scripts]
api-gateway = "src.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
